智能合约

+++ tayi/src/main/java/com/xymtop/tayi/core/graph/Neo4jUtils.java
@@ -4,7 +4,9 @@ import cn.hutool.core.bean.BeanUtil;
 import cn.hutool.core.map.MapUtil;
 import com.xymtop.tayi.core.graph.entity.NodeRelationship;
 import com.xymtop.tayi.core.nft.NFTData;
+import com.xymtop.tayi.core.nft.NFTMeta;
 import lombok.Data;
+import org.neo4j.cypher.internal.expressions.functions.E;
 import org.neo4j.dbms.api.DatabaseManagementService;
 import org.neo4j.dbms.api.DatabaseManagementServiceBuilder;
 import org.neo4j.graphdb.*;
@@ -50,6 +52,12 @@ public class Neo4jUtils {
         try (Transaction tx = graphDb.beginTx()) {
             Node startNode = tx.findNode(Label.label(NFTUtils.LABEL), "address", startNodeId);
             Node endNode = tx.findNode(Label.label(NFTUtils.LABEL), "address", endNodeId);
+
+
+            if (startNode == null){
+                return null;
+            }
+
             RelationshipType type = RelationshipType.withName(relation);
             Relationship relationship = startNode.createRelationshipTo(endNode, type);
             tx.commit();
@@ -66,25 +74,50 @@ public class Neo4jUtils {
                 relationship.delete();
             }
             Node node = tx.findNode(Label.label(NFTUtils.LABEL), "address", nftAddress);
-            node.delete();
+            if (node!=null){
+                node.delete();
+            }
+
             tx.commit();
         }
         return true;
     }
 
 
+
+    public NFTData beanToNftData(Map<String, Object> map){
+        Map<String,Object> newMap = new HashMap<>();
+        for (String key : map.keySet()){
+            if (key.startsWith("nft-")){
+                String newKey = key.split("nft-")[1];
+                newMap.put(newKey,map.get(key));
+            }
+        }
+
+        NFTMeta meta = new NFTMeta();
+       meta.setAttributes(newMap);
+
+        NFTData nftData = new NFTData();
+        nftData =  BeanUtil.mapToBean(map,NFTData.class,true);
+        nftData.setMeta(meta);
+
+        return nftData;
+    }
+
     //获取到NFT数据
     public NFTData getById(String address){
         try (Transaction tx = graphDb.beginTx()) {
             Node node = tx.findNode(Label.label(NFTUtils.LABEL), "address", address);
+            if (node == null){
+                return null;
+            }
 
             Map<String, Object> allProperties = node.getAllProperties();
             NFTData nftData = new NFTData();
 
             //转对象
-            nftData = BeanUtil.mapToBean(allProperties, NFTData.class, true);
+            nftData = beanToNftData(allProperties);
 
-            node.delete();
             tx.commit();
             return nftData;
         }
@@ -108,6 +141,21 @@ public class Neo4jUtils {
         }
     }
 
+    public boolean updateNftNodeProperties(String nftId,NFTData nftData) {
+        Map<String, Object> properties = BeanUtil.beanToMap(nftData);
+        try (Transaction tx = graphDb.beginTx()) {
+            Node node = tx.findNode(Label.label(NFTUtils.LABEL), "nftId", nftId);
+
+            if (node == null){
+                return false;
+            }
+            properties.forEach(node::setProperty);
+            tx.commit();
+            return true;
+        }catch (Exception e){
+            return false;
+        }
+    }
     public void updateNodeProperties(Node node, Map<String, Object> properties) {
         try (Transaction tx = graphDb.beginTx()) {
             properties.forEach(node::setProperty);
@@ -119,6 +167,36 @@ public class Neo4jUtils {
         try (Transaction tx = graphDb.beginTx()) {
             relationship.delete();
             tx.commit();
+        }
+    }
+
+    public boolean deleteNftRelationship(String nftId1, String nftId2,String relationshipType){
+        try (Transaction tx = graphDb.beginTx()) {
+
+            Node node1 = tx.findNode(Label.label(NFTUtils.LABEL), "nftId", nftId1);
+            Node node2 = tx.findNode(Label.label(NFTUtils.LABEL), "nftId", nftId2);
+
+            if (node1 == null){
+                return false;
+            }
+
+            if (node2 == null){
+                return false;
+            }
+            RelationshipType relationshipTypeTemp = RelationshipType.withName(relationshipType);
+
+            ResourceIterable<Relationship> relationships = node1.getRelationships();
+            for (Relationship relationship : relationships){
+                if (relationship.getType().equals(relationshipTypeTemp) && relationship.getOtherNode(node1).equals(node2)){
+                    relationship.delete();
+                    break;
+                }
+            }
+
+            tx.commit();
+            return true;
+        }catch (Exception e){
+            return false;
         }
     }
 
@@ -162,11 +240,22 @@ public class Neo4jUtils {
 
         try (Transaction tx = graphDb.beginTx()) {
 
-            ResourceIterable<Relationship> allRelationships = tx.getAllRelationships();
+            Node node = tx.findNode(Label.label(NFTUtils.LABEL), "address", nftId);
+
+            if (node == null){
+                return null;
+            }
+            ResourceIterable<Relationship> allRelationships = node.getRelationships();
             for (Relationship relationship : allRelationships){
                 Node endNode = relationship.getEndNode();
-                NFTData nftData = nodeToNFT(endNode);
-                nodes.add(nftData);
+                Node startNode = relationship.getStartNode();
+                if (node == startNode){
+                    NFTData nftData = nodeToNFT(endNode);
+                    nodes.add(nftData);
+                }else {
+                    NFTData nftData = nodeToNFT(startNode);
+                    nodes.add(nftData);
+                }
             }
             tx.commit();
             return nodes;
@@ -177,7 +266,13 @@ public class Neo4jUtils {
         try (Transaction tx = graphDb.beginTx()) {
             List<NodeRelationship> resultList = new ArrayList<>();
 
-            ResourceIterable<Relationship> allRelationships = tx.getAllRelationships();
+            Node node = tx.findNode(Label.label(NFTUtils.LABEL), "address", nftId);
+
+            if (node == null){
+                return null;
+            }
+            ResourceIterable<Relationship> allRelationships = node.getRelationships();
+
             for (Relationship relationship : allRelationships){
                 RelationshipType type = relationship.getType();
                 Node startNode = relationship.getStartNode();
@@ -186,7 +281,7 @@ public class Neo4jUtils {
                 NFTData nftDataEnd = nodeToNFT(endNode);
 
                 NodeRelationship nodeRelationship = new NodeRelationship();
-                nodeRelationship.setEndNode(nftDataEnd);
+                nodeRelationship.setStartNode(nftDataStart);
                 nodeRelationship.setEndNode(nftDataEnd);
                 nodeRelationship.setType(type.name());
 
@@ -204,7 +299,7 @@ public class Neo4jUtils {
         NFTData nftData = new NFTData();
 
         //转对象
-        nftData = BeanUtil.mapToBean(allProperties, NFTData.class, true);
+        nftData = beanToNftData(allProperties);
 
         return nftData;
     }
@@ -279,6 +374,41 @@ public class Neo4jUtils {
         try (Transaction tx = graphDb.beginTx()) {
             tx.execute("MATCH (n) DETACH DELETE n");
             tx.commit();
+        }
+    }
+
+
+    //根据对象来查询NFT
+    public List<NFTData> queryNFTByObject(NFTData nftData) {
+        List<NFTData> nftDataList = new ArrayList<>();
+        try (Transaction tx = graphDb.beginTx()) {
+            Map<String, Object> properties = NFTUtils.createNFTMap(nftData);
+
+            // 构建匹配条件
+            StringBuilder matchCondition = new StringBuilder();
+            for (String key : properties.keySet()) {
+                matchCondition.append(String.format("n.%s = $%s AND ", key, key));
+            }
+
+            // 移除最后的 "AND "
+            String conditions = matchCondition.substring(0, matchCondition.length() - 5);
+
+            // 构建查询
+            String query = String.format("MATCH (n:%s) WHERE %s RETURN n",NFTUtils.LABEL, conditions);
+
+            // 执行查询
+            Result result = tx.execute(query, properties);
+            while (result.hasNext()) {
+                Node node = (Node) result.next().get("n");
+
+                NFTData data = nodeToNFT(node);
+
+                nftDataList.add(data);
+            }
+            tx.commit();
+            return nftDataList;
+        }catch (Exception e){
+            return null;
         }
     }
